+++ packages/euclid/src/msgs/pool_factory/mod.rs
@@ -1,3 +1,5 @@
 pub mod msg;
+pub mod reply;
 
 pub use msg::*;
+pub use reply::*;